moves all field component prop types to payload
