+++ packages/payload/src/admin/fields/Code.ts
@@ -1,0 +1,10 @@
+import type { CodeField } from '../../fields/config/types.js'
+import type { FormFieldBase } from '../types.js'
+
+export type CodeFieldProps = {
+  editorOptions?: CodeField['admin']['editorOptions']
+  language?: CodeField['admin']['language']
+  name?: string
+  path?: string
+  width: string
+} & FormFieldBase
