+++ packages/payload/src/admin/fields/Collapsible.ts
@@ -1,0 +1,8 @@
+import type { FieldMap } from '../forms/FieldMap.js'
+import type { FormFieldBase } from '../types.js'
+
+export type CollapsibleFieldProps = {
+  fieldMap: FieldMap
+  initCollapsed?: boolean
+  width?: string
+} & FormFieldBase
