+++ packages/payload/src/admin/fields/Date.ts
@@ -1,0 +1,10 @@
+import type { DateField } from '../../fields/config/types.js'
+import type { FormFieldBase } from '../types.js'
+
+export type DateFieldProps = {
+  date?: DateField['admin']['date']
+  name?: string
+  path?: string
+  placeholder?: DateField['admin']['placeholder'] | string
+  width?: string
+} & FormFieldBase
