+++ packages/payload/src/admin/fields/Email.ts
@@ -1,0 +1,10 @@
+import type { EmailField } from '../../fields/config/types.js'
+import type { FormFieldBase } from '../types.js'
+
+export type EmailFieldProps = {
+  autoComplete?: string
+  name?: string
+  path?: string
+  placeholder?: EmailField['admin']['placeholder']
+  width?: string
+} & FormFieldBase
