+++ packages/payload/src/admin/fields/Group.ts
@@ -1,0 +1,10 @@
+import type { FieldMap } from '../forms/FieldMap.js'
+import type { FormFieldBase } from '../types.js'
+
+export type GroupFieldProps = {
+  fieldMap: FieldMap
+  forceRender?: boolean
+  hideGutter?: boolean
+  name?: string
+  width?: string
+} & FormFieldBase
