+++ packages/payload/src/admin/fields/Hidden.ts
@@ -1,0 +1,9 @@
+import type { FormFieldBase } from '../types.js'
+
+export type HiddenFieldProps = {
+  disableModifyingForm?: false
+  forceUsePathFromProps?: boolean
+  name?: string
+  path?: string
+  value?: unknown
+} & FormFieldBase
