+++ packages/payload/src/admin/fields/Point.ts
@@ -1,0 +1,9 @@
+import type { FormFieldBase } from '../types.js'
+
+export type PointFieldProps = {
+  name?: string
+  path?: string
+  placeholder?: string
+  step?: number
+  width?: string
+} & FormFieldBase
