+++ packages/payload/src/admin/fields/Relationship.ts
@@ -1,0 +1,12 @@
+import type { RelationshipField } from '../../fields/config/types.js'
+import type { FormFieldBase } from '../types.js'
+
+export type RelationshipFieldProps = {
+  allowCreate?: RelationshipField['admin']['allowCreate']
+  hasMany?: boolean
+  isSortable?: boolean
+  name: string
+  relationTo?: RelationshipField['relationTo']
+  sortOptions?: RelationshipField['admin']['sortOptions']
+  width?: string
+} & FormFieldBase
