+++ packages/payload/src/admin/fields/RichText.ts
@@ -1,0 +1,8 @@
+import type { MappedField } from '../forms/FieldMap.js'
+import type { FormFieldBase } from '../types.js'
+
+export type RichTextComponentProps = {
+  name: string
+  richTextComponentMap?: Map<string, MappedField[] | React.ReactNode>
+  width?: string
+} & FormFieldBase
