+++ packages/payload/src/admin/fields/Text.ts
@@ -1,0 +1,16 @@
+import type { TextField } from '../../fields/config/types.js'
+import type { FormFieldBase } from '../types.js'
+
+export type TextFieldProps = {
+  hasMany?: boolean
+  inputRef?: React.MutableRefObject<HTMLInputElement>
+  maxLength?: number
+  maxRows?: number
+  minLength?: number
+  minRows?: number
+  name?: string
+  onKeyDown?: React.KeyboardEventHandler<HTMLInputElement>
+  path?: string
+  placeholder?: TextField['admin']['placeholder']
+  width?: string
+} & FormFieldBase
