+++ packages/ui/src/fields/Array/index.tsx
@@ -1,10 +1,8 @@
 'use client'
-import type { ArrayField as ArrayFieldType, FormFieldBase } from 'payload'
+import type { ArrayFieldProps, ArrayField as ArrayFieldType } from 'payload'
 
 import { getTranslation } from '@payloadcms/translations'
 import React, { useCallback } from 'react'
-
-import type { FieldMap } from '../../providers/ComponentMap/buildComponentMap/types.js'
 
 import { Banner } from '../../elements/Banner/index.js'
 import { Button } from '../../elements/Button/index.js'
@@ -29,18 +27,6 @@ import { ArrayRow } from './ArrayRow.js'
 import './index.scss'
 
 const baseClass = 'array-field'
-
-export type ArrayFieldProps = {
-  CustomRowLabel?: React.ReactNode
-  fieldMap: FieldMap
-  forceRender?: boolean
-  isSortable?: boolean
-  labels?: ArrayFieldType['labels']
-  maxRows?: ArrayFieldType['maxRows']
-  minRows?: ArrayFieldType['minRows']
-  name?: string
-  width?: string
-} & FormFieldBase
 
 export const _ArrayField: React.FC<ArrayFieldProps> = (props) => {
   const {
@@ -72,6 +58,7 @@ export const _ArrayField: React.FC<ArrayFieldProps> = (props) => {
     permissions,
     readOnly: readOnlyFromContext,
   } = useFieldProps()
+
   const minRows = minRowsProp ?? required ? 1 : 0
 
   const { setDocFieldPreferences } = useDocumentInfo()
