+++ packages/ui/src/fields/Checkbox/index.tsx
@@ -1,10 +1,9 @@
 'use client'
-import type { ClientValidate } from 'payload'
+import type { ClientValidate, CheckboxFieldProps } from 'payload'
 
 import React, { useCallback } from 'react'
 
 import type { CheckboxInputProps } from './Input.js'
-import type { CheckboxFieldProps } from './types.js'
 
 import { useFieldProps } from '../../forms/FieldPropsProvider/index.js'
 import { useForm } from '../../forms/Form/context.js'
