+++ packages/ui/src/fields/Code/index.tsx
@@ -1,5 +1,5 @@
 'use client'
-import type { CodeField as CodeFieldType, FormFieldBase } from 'payload'
+import type { CodeFieldProps } from 'payload'
 
 import React, { useCallback } from 'react'
 
@@ -12,14 +12,6 @@ import { FieldError } from '../FieldError/index.js'
 import { FieldLabel } from '../FieldLabel/index.js'
 import { fieldBaseClass } from '../shared/index.js'
 import './index.scss'
-
-export type CodeFieldProps = {
-  editorOptions?: CodeFieldType['admin']['editorOptions']
-  language?: CodeFieldType['admin']['language']
-  name?: string
-  path?: string
-  width: string
-} & FormFieldBase
 
 const prismToMonacoLanguageMap = {
   js: 'javascript',
