+++ packages/ui/src/fields/JSON/index.tsx
@@ -1,5 +1,5 @@
 'use client'
-import type { ClientValidate, FormFieldBase, JSONField as JSONFieldType } from 'payload'
+import type { ClientValidate, JSONFieldProps } from 'payload'
 
 import React, { useCallback, useEffect, useState } from 'react'
 
@@ -15,14 +15,6 @@ const baseClass = 'json-field'
 import { useFieldProps } from '../../forms/FieldPropsProvider/index.js'
 import { FieldDescription } from '../FieldDescription/index.js'
 import { FieldError } from '../FieldError/index.js'
-
-export type JSONFieldProps = {
-  editorOptions?: JSONFieldType['admin']['editorOptions']
-  jsonSchema?: Record<string, unknown>
-  name?: string
-  path?: string
-  width?: string
-} & FormFieldBase
 
 const _JSONField: React.FC<JSONFieldProps> = (props) => {
   const {
