+++ packages/ui/src/fields/Relationship/types.ts
@@ -1,20 +1,5 @@
 import type { I18nClient } from '@payloadcms/translations'
-import type {
-  ClientCollectionConfig,
-  FormFieldBase,
-  RelationshipField,
-  SanitizedConfig,
-} from 'payload'
-
-export type RelationshipFieldProps = {
-  allowCreate?: RelationshipField['admin']['allowCreate']
-  hasMany?: boolean
-  isSortable?: boolean
-  name: string
-  relationTo?: RelationshipField['relationTo']
-  sortOptions?: RelationshipField['admin']['sortOptions']
-  width?: string
-} & FormFieldBase
+import type { ClientCollectionConfig, SanitizedConfig } from 'payload'
 
 export type Option = {
   label: string
