+++ packages/ui/src/fields/Row/index.tsx
@@ -1,14 +1,13 @@
 'use client'
 import React from 'react'
 
-import type { RowFieldProps } from './types.js'
-
 import { useFieldProps } from '../../forms/FieldPropsProvider/index.js'
 import { RenderFields } from '../../forms/RenderFields/index.js'
 import { withCondition } from '../../forms/withCondition/index.js'
 import { fieldBaseClass } from '../shared/index.js'
 import './index.scss'
 import { RowProvider, useRow } from './provider.js'
+import type { RowFieldProps } from 'payload'
 
 export { RowProvider, useRow }
