+++ packages/ui/src/fields/Select/index.tsx
@@ -1,5 +1,5 @@
 'use client'
-import type { ClientValidate, FormFieldBase, Option, OptionObject } from 'payload'
+import type { ClientValidate, Option, OptionObject, SelectFieldProps } from 'payload'
 
 import React, { useCallback } from 'react'
 
@@ -10,18 +10,6 @@ import { useFieldProps } from '../../forms/FieldPropsProvider/index.js'
 import { useField } from '../../forms/useField/index.js'
 import { withCondition } from '../../forms/withCondition/index.js'
 import { SelectInput } from './Input.js'
-
-export type SelectFieldProps = {
-  hasMany?: boolean
-  isClearable?: boolean
-  isSortable?: boolean
-  name?: string
-  onChange?: (e: string | string[]) => void
-  options?: Option[]
-  path?: string
-  value?: string
-  width?: string
-} & FormFieldBase
 
 const formatOptions = (options: Option[]): OptionObject[] =>
   options.map((option) => {
