+++ packages/ui/src/fields/Text/types.ts
@@ -1,21 +1,7 @@
-import type { FormFieldBase, TextField } from 'payload'
+import type { TextFieldProps } from 'payload'
 import type { ChangeEvent } from 'react'
 
 import type { Option, ReactSelectAdapterProps } from '../../elements/ReactSelect/types.js'
-
-export type TextFieldProps = {
-  hasMany?: boolean
-  inputRef?: React.MutableRefObject<HTMLInputElement>
-  maxLength?: number
-  maxRows?: number
-  minLength?: number
-  minRows?: number
-  name?: string
-  onKeyDown?: React.KeyboardEventHandler<HTMLInputElement>
-  path?: string
-  placeholder?: TextField['admin']['placeholder']
-  width?: string
-} & FormFieldBase
 
 export type SharedTextFieldProps =
   | {
