+++ packages/ui/src/fields/Textarea/types.ts
@@ -1,16 +1,5 @@
-import type { FormFieldBase, TextareaField as TextareaFieldType } from 'payload'
-
+import type { TextareaFieldProps } from 'packages/payload/src/index.js'
 import { type ChangeEvent } from 'react'
-
-export type TextareaFieldProps = {
-  maxLength?: number
-  minLength?: number
-  name?: string
-  path?: string
-  placeholder?: TextareaFieldType['admin']['placeholder']
-  rows?: number
-  width?: string
-} & FormFieldBase
 
 export type TextAreaInputProps = {
   onChange?: (e: ChangeEvent<HTMLTextAreaElement>) => void
